+++ src/application/article/get.rs
@@ -38,7 +38,7 @@ pub struct GetArticleOutput {
     pub content: String, // todo: server rendering 
     pub state: ArticleState,
     pub views: u32,
-    pub rating: i32,
+    pub rating: i64,
     pub author_id: UserId,
     pub tags: Vec<Tag>,
